some sanity checking before doing a kill
Sun Jul 07 06:26:51 2002  George Lebl <jirka@5z.com>

	* gui/gdmsetup.c: some sanity checking before doing a kill

+++ gui/gdmsetup.c
@@ -83,6 +83,11 @@ update_greeters (void)
 			g_free (ret);
 			goto check_update_error;
 		}
+
+		/* sanity */
+		if (pid <= 0)
+			continue;
+
 		if (kill (pid, SIGHUP) != 0)
 			have_error = TRUE;
 		p = strchr (p, ';');
